+++ app/src/main/java/com/concordium/wallet/ui/cis2/TokensFragment.kt
@@ -4,10 +4,14 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.content.res.AppCompatResources
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.concordium.wallet.R
 import com.concordium.wallet.data.model.Token
 import com.concordium.wallet.databinding.FragmentTokensBinding
+
 
 class TokensFragment : Fragment() {
     private var _binding: FragmentTokensBinding? = null
@@ -19,14 +23,19 @@ class TokensFragment : Fragment() {
 
     companion object {
         @JvmStatic
-        fun newInstance(viewModel: TokensViewModel, accountAddress: String, isFungible: Boolean) = TokensFragment().apply {
-            _viewModel = viewModel
-            _accountAddress = accountAddress
-            _isFungible = isFungible
-        }
+        fun newInstance(viewModel: TokensViewModel, accountAddress: String, isFungible: Boolean) =
+            TokensFragment().apply {
+                _viewModel = viewModel
+                _accountAddress = accountAddress
+                _isFungible = isFungible
+            }
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
         _binding = FragmentTokensBinding.inflate(inflater, container, false)
         return binding.root
     }
@@ -39,16 +48,32 @@ class TokensFragment : Fragment() {
     }
 
     private fun initViews() {
-        binding.tokensFound.layoutManager = LinearLayoutManager(activity)
-        tokensAccountDetailsAdapter = TokensAccountDetailsAdapter(requireContext(), _isFungible, _isFungible, arrayOf())
+        val layoutManager = LinearLayoutManager(activity)
+        val dividerItemDecoration = DividerItemDecoration(
+            requireContext(),
+            layoutManager.orientation
+        )
+        binding.tokensFound.addItemDecoration(
+            TokenItemSeparatorDecorator(
+                AppCompatResources.getDrawable(
+                    requireContext(),
+                    R.drawable.divider
+                )
+            )
+        )
+        binding.tokensFound.layoutManager = layoutManager
+
+        tokensAccountDetailsAdapter =
+            TokensAccountDetailsAdapter(requireContext(), _isFungible, true, arrayOf())
         tokensAccountDetailsAdapter.also { binding.tokensFound.adapter = it }
 
-        tokensAccountDetailsAdapter.setTokenClickListener(object : TokensAccountDetailsAdapter.TokenClickListener {
+        tokensAccountDetailsAdapter.setTokenClickListener(object :
+            TokensAccountDetailsAdapter.TokenClickListener {
             override fun onRowClick(token: Token) {
                 _viewModel.chooseToken.postValue(token)
             }
-            override fun onCheckBoxClick(token: Token) {
-            }
+
+            override fun onCheckBoxClick(token: Token) = Unit
         })
     }